pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "sqlite-utils>=3.39",
    "streamlit>=1.55.0",
]
